+++ src/App.js
@@ -2,12 +2,13 @@ import React from 'react';
 import { Router, Route } from 'react-router-dom';
 import { connect } from 'react-redux'
 import ReactGA from 'react-ga';
-import './App.css';
-import Login from './components/Login';
-import Logout from './components/Logout';
 import history from './history';
 import AuthCallback from './components/AuthCallback';
+import Home from './views/Home';
+import Location from './views/Location';
 import config from './config';
+import Button from 'react-bootstrap/Button';
+import Header from './components/Header';
 
 function App(props) {
 
@@ -17,20 +18,14 @@ function App(props) {
   ReactGA.pageview(window.location.pathname + window.location.search);
 
   return (
-    <Router history={history}>
-      <Route exact path="/" render={() => (
-        <div className="App">
-          <header className="App-header">
-            {
-              !props.auth.isAuthenticated ?
-              <Login/>
-              : <Logout/>
-            }
-          </header>
-        </div>
-      )} />
-      <Route exact path="/auth-callback" component={AuthCallback} />
-    </Router>
+    <div>
+      <Header/>
+      <Router history={history}>
+        <Route exact path="/" component={Home} />
+        <Route exact path="/auth-callback" component={AuthCallback} />
+        <Route exact path="/l/:code" component={Location} />
+      </Router>
+    </div>
   );
 }