+++ CHANGELOG.md
@@ -1,7 +1,17 @@
-`0.1.0-dev`
+# Change Log
 
+`0.2.0-dev`
 
+---
+`0.2.0`
+# Minor Changes
+ - Add React Bootstrap
+ - Plus Codes Integration
+ 
+# Patches
+ - Netlify redirects file to fix auth callback
 
+---
 `0.1.0`
 
 # Minor Changes